backend/requirements.txt
django>=4.2
psycopg2-binary
python-dotenv
asgiref
uvicorn
djangorestframework
django-cors-headers
pyjwt
djangorestframework-simplejwt